api-mongo/config/database.py
from pymongo import MongoClient
from dotenv import load_dotenv
from typing import Generator
import os

load_dotenv()

MONGO_URI = os.getenv("MONGO_URI")
MONGO_DB = os.getenv("MONGO_DB")


def get_mongo_client():
    return MongoClient(MONGO_URI)


def get_database():
    client = get_mongo_client()
    return client[MONGO_DB]


db = get_database()