api-mongo/pyproject.toml
[project]
name = "api-mongo"
version = "0.1.0"
description = "Backend service to create orders of products"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "pydantic-settings>=2.0.0",
    "pymongo>=4.0.0",
    "python-dotenv>=1.0.0",
]
[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",  # needed for FastAPI's TestClient
]